chore: domain config 통합
